requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docx-mailmerge==0.5.0
Flask==1.1.0
Flask-Cors==3.0.8
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
